Moved: storage service to another package
