+++ internal/service/module.go
@@ -3,12 +3,13 @@ package service
 import (
 	"go.uber.org/fx"
 
+	"github.com/gomvn/gomvn/internal/service/storage"
 	"github.com/gomvn/gomvn/internal/service/user"
 )
 
 var Module = fx.Options(
 	fx.Provide(NewPathService),
-	fx.Provide(NewLocalStorage),
+	fx.Provide(storage.NewLocalStorage),
 	fx.Provide(NewRepoService),
 	fx.Provide(user.New),
 	fx.Invoke(user.Initialize),
